Moved DataUtil NatVsTech and RFECVResult from root to src/nanogbcdt, and removed old source files
